refactor: reorganize file structure a little bit
